Working on splitting the MVC
