+++ index.html
@@ -11,18 +11,16 @@
     <main>
       <section class="game">
         <section class="status">
-          <div class="round-counter" id="round-counter">
-            # of 12
-          </div>
+          <div class="round-counter" id="round-counter"># of 12</div>
         </section>
         <section class="board">
           <!-- <div id="dev-solve"></div> -->
-          <div id="rounds"></div><div id="score-box"></div>
-          <div class="colour-options" id="choiceBox">
+          <div class="rounds-board" id="roundsBoard">
+            <div class="rounds-round" id="rounds"></div>
+            <div class="rounds-score" id="score-box"></div>
           </div>
-          <button class="submit" name="submit" type="button" id="submit">
-            Submit Round
-          </button>
+          <div class="colour-options" id="choiceBox"></div>
+          <div class="button-field" id="buttonField"></div>
         </section>
       </section>
     </main>
